requirements.txt
torch
torchvision
transformers
timm
einops
supervision==0.18.0
opencv-python
pillow
numpy
scipy
tqdm